Split DPM upstream client protocols
