+++ src/app/services/upstream_client_protocols.py
@@ -1,5 +1,12 @@
 from typing import Any, Protocol
 
+from app.services.dpm_client_protocols import (
+    DpmCommandCenterClient as DpmCommandCenterClient,
+)
+from app.services.dpm_client_protocols import (
+    DpmWaveClient as DpmWaveClient,
+)
+
 
 class DpmConstructionClient(Protocol):
     async def generate_construction_alternative_set(
@@ -1454,544 +1461,3 @@ class AdvisorBriefAdviseClient(Protocol):
         correlation_id: str,
     ) -> tuple[int, dict[str, Any]]: ...
 
-
-class DpmCommandCenterClient(Protocol):
-    async def get_command_center(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def run_monitoring_once(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_monitoring_runs(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_monitoring_run(
-        self,
-        monitoring_run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_monitoring_exceptions(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def resolve_monitoring_exception(
-        self,
-        exception_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_mandate_by_portfolio(
-        self,
-        portfolio_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_mandate(
-        self,
-        mandate_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_mandate_health(
-        self,
-        mandate_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_mandate_diff(
-        self,
-        mandate_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def preview_outcome_review(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_outcome_review(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_outcome_reviews(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_outcome_review(
-        self,
-        outcome_review_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def refresh_outcome_review_sources(
-        self,
-        outcome_review_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_outcome_review_supportability(
-        self,
-        outcome_review_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_outcome_review_report_input(
-        self,
-        outcome_review_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_outcome_review_ai_evidence_input(
-        self,
-        outcome_review_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_run_outcome_review(
-        self,
-        rebalance_run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_wave_outcome_reviews(
-        self,
-        wave_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_portfolio_memory(
-        self,
-        portfolio_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def search_portfolio_memory(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def preview_pm_operating_quality_score_run(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_pm_operating_quality_score_run(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_pm_operating_quality_score_runs(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_pm_operating_quality_score_run(
-        self,
-        score_run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def preview_pm_operating_quality_summary_invocation(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_pm_operating_quality_summary_invocation(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_pm_operating_quality_summary_invocations(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_pm_operating_quality_summary_invocation(
-        self,
-        summary_invocation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def preview_pm_operating_quality_review_action(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_pm_operating_quality_review_action(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_pm_operating_quality_review_actions(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_pm_operating_quality_review_action(
-        self,
-        review_action_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def preview_pm_operating_quality_fairness_analysis(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_pm_operating_quality_fairness_analysis(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_pm_operating_quality_fairness_analyses(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_pm_operating_quality_fairness_analysis(
-        self,
-        fairness_analysis_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def put_pm_operating_quality_policy(
-        self,
-        policy_id: str,
-        policy_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_pm_operating_quality_policies(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_pm_operating_quality_policy(
-        self,
-        policy_id: str,
-        policy_version: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class DpmWaveClient(Protocol):
-    async def preview_wave(
-        self,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_wave(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_waves(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_wave(
-        self,
-        wave_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def put_campaign_definition(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_campaign_definitions(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_definition(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_definition_lifecycle_events(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_definition_preview_readiness(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_definition_launch_history(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_definition_launch_package(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def launch_campaign_definition(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def retire_campaign_definition(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def supersede_campaign_definition(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def discover_campaigns(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_operating_queue(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_approval_inbox(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_workflow_board(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_assignment_plan(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_campaign_workflow_automation(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_campaign_approval_decisions(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_campaign_approval_decision(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_campaign_assignment_actions(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_campaign_assignment_action(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_campaign_assignment_tasks(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_campaign_assignment_task(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def transition_campaign_assignment_task(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        task_ref: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_campaign_maker_checker_controls(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_campaign_maker_checker_control(
-        self,
-        campaign_id: str,
-        campaign_version: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_wave_items(
-        self,
-        wave_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def source_check_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def simulate_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def select_wave_item(
-        self,
-        wave_id: str,
-        wave_item_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def approve_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def stage_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def handoff_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def cancel_wave(
-        self,
-        wave_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_wave_proof_pack_posture(
-        self,
-        wave_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_wave_supportability(
-        self,
-        wave_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_wave_report_input(
-        self,
-        wave_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
